Implement APIs for texture update in a non-GL thread.
Several changes included.

1. Introduced GVRFutureOnGlThread. This is a special implementation of RunnableFuture.
This should be used when you want a return value and run it with GVRContext.runOnGlThread(Runnable).

2. Add new APIs safeUpdate with GVRBitmapTexture. Previously GVRBitmapTexture.update
running on a non-GL thread won't return or report any error but you can't see your
update on screen. Using safeUpdate will make any update call from a non-GL thread
successful on screen.

3. Add new API GVRContext.isCurrentThreadGLThread. Previously, we only have API
assertGLThread inside GVRF to check if current thread is GL thread or throws an
Exception. I think it should be fine to open an API to check if current thread is
GL thread to judge if we should directly call anything that requires GL thread and
to see if we should use GVRFutureOnGlThread.

4. Fix GVRConsole. Previously, if you want to call GVRConsole.writeLine in a non-GL
thread, ex, through an input event, it won't update anything on screen. With
GVRBitmapTexture.safeUpdate, this is fixed.