gvr-cubemap: Add a new sample to show efficient cube mapping
Currenty, the samples show how to construct a cube scene object using a list of
textures.

A sample is added to use a GVRCubemapTexture which can be loaded from a .zip file.

Using the cubemap, each frame only requires 1 draw and 12 triangles. It is also more
convnient than using a list of textures.