Fix offset issue in compressed textures
When ByteBuffer wraps a subarray of byte[], ByteBuffer.array() returns
the original backing array without the offset. This gives a wrong pointer to
the texture data.

CompressedTexture is modified to obtain the offset as ByteBuffer.position()
before reading from it, and exposes methods getArray(), getArrayOffset(),
and getData() to retrieve the backing array, data offset in backing array, and the
ByteBuffer itself, respectively.

GearVRf-DCO-1.0-Signed-off-by: Danke Xie
d.xie@samsung.com