Fix JNI crash due to using a JNI environment attached to the UI thread from the Oculus native VR thread on shutdown. Plus small cleanup.
Unfortunately there is another crash, this time in Oculus so this
does't help. Not sure I will be able to do something about the Oculus
crash (in vrapi_Shutdown there seems to be use of an invalid JNI
object). The Oculus VrCubeWorld_Framework sample exhibits the same
behavior..

So this is it for now. Will try few more things but we might have to
hope that the new Oculus SDK remedies the problem.