Fix for issue #286, "Textures not loading while loading a 3d model"
Reduces the number of threads needed while loading a model. setMainTexture
spawns a thread that waits for the texture to load. But if you do this
for enough textures at the right time, all the threads can be used up and
there will be no thread available to perform the actual loading..

The problem is still there and if someone who is overeager to use
Future<Texture>s can experience it. I will look at it again later and
might have to introduce one more pool.