Fix for bug 19478, "Crash while launching an app"
Didn't reproduce it thus far; this is my best guess.

The thread-safety guarantee on the Java side is enough but the
changes to the C++ are nice-to-have  - if ever that code is used
without relying on the Java side for locking, these checks are
necessary as nullptr is valid return value.

Change-Id: I9bd6a00d118005148607a025c13b2ed513ca9b61