handle scene switching on collider list
Added Scene::set_main_scene to let C++ layer know what is the mai scene.
Colliders are added to the main scene's collider list (which may not
always be correct). When the scene is switched, the colliders are
gathered from the hierarchy, a one-time overhead. Thereafter they are
added to this scene.